combined server and user commands under info
